Remove vfReachable and modify IsReachable to only use vfLimited.
This is a backport from PIVX commit 38ed737616076f44dcdb193cf2687c54a04c2b15
This should prevent some net related issues. (including the invalid IP
related errors when attempting to start a correctly configured masternode
from the gui)

Original comment:
This is a backport from bitcoin commit 110b62f
This should fix issue 373 and prevent some future net related issues
Original comment:
We do not know that a class of Network is reachable, only that it is not.